refactor: extract UBThumbnailTextItem
- UBThumbnailTextItem was part of UBDocumentThumbnailsView
- extract to own pair of files to improve code structure
