+++ src/gui/gui.pri
@@ -2,6 +2,7 @@ HEADERS +=  \
     src/gui/UBThumbnail.h \
     src/gui/UBThumbnailArranger.h \
     src/gui/UBThumbnailScene.h \
+    src/gui/UBThumbnailTextItem.h \
     src/gui/UBThumbnailsView.h \
     $$PWD/UBStartupHintsPalette.h \
     src/gui/UBFloatingPalette.h \
@@ -50,6 +51,7 @@ SOURCES +=  \
     src/gui/UBThumbnail.cpp \
     src/gui/UBThumbnailArranger.cpp \
     src/gui/UBThumbnailScene.cpp \
+    src/gui/UBThumbnailTextItem.cpp \
     src/gui/UBThumbnailsView.cpp \
     $$PWD/UBStartupHintsPalette.cpp \
     src/gui/UBFloatingPalette.cpp \
